pyproject.toml
[project]
name = "hallucinot"
version = "0.1.0"
description = "A simple citation-checking app for legal documents."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "eyecite==2.7.6",
  "flask==3.1.3",
  "pypdf==6.9.2",
  "python-docx==1.2.0",
  "requests==2.32.5",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["hallucinot*"]
